refactor automl predictor
