+++ pyzoo/zoo/automl/search/RayTuneSearchEngine.py
@@ -56,8 +56,6 @@ class RayTuneSearchEngine(SearchEngine):
                 search_space,
                 recipe,
                 feature_transformers=None,
-                # model=None,
-                future_seq_len=1,
                 validation_df=None,
                 mc=False,
                 metric="mse",
@@ -132,7 +130,6 @@ class RayTuneSearchEngine(SearchEngine):
         self.train_func = self._prepare_train_func(input_df=input_df,
                                                    model_create_func=model_create_func,
                                                    feature_transformers=feature_transformers,
-                                                   future_seq_len=future_seq_len,
                                                    validation_df=validation_df,
                                                    metric=metric,
                                                    metric_mode=metric_mode,
@@ -239,7 +236,6 @@ class RayTuneSearchEngine(SearchEngine):
     def _prepare_train_func(input_df,
                             model_create_func,
                             feature_transformers,
-                            future_seq_len,
                             metric,
                             metric_mode,
                             validation_df=None,
